requirements.txt
aniso8601==7.0.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
gunicorn==19.9.0
heroku==0.1.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.0
pandas==0.25.0
pyenchant==1.6.6
pyparsing==2.4.2
python-dateutil==1.5
pytz==2019.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
